Refactored ChartBuilder.cs
Now stores the kdaChartLists as a list of lists inside of azure blob storage. This is then loaded when InitChartData is called so that the next iteration of data can be added to the kdaChartLists.

Also store chartData as a dictionary within azure blob storage so ChartFunction.cs can load and build chartURL